requirements.txt
django>=5.0,<6.0
requests>=2.31
whitenoise>=6.6.0
gunicorn>=21.2.0
django-redis>=5.4.0
hiredis>=2.3.2
uvicorn[standard]>=0.29.0
httpx>=0.27.0
